splitted the header and source
