+++ include/RFID/RFID.hpp
@@ -2,6 +2,7 @@
 
 #include <Arduino.h>
 #include <MFRC522.h>
+#include <Adafruit_PN532.h>
 #include "../Unlock_Object.hpp"
 #include "../Helper.hpp"
 #include "UID.hpp"
@@ -23,17 +24,22 @@ namespace RFID
     class RFID : public Unlock_Object
     {
     public:
-        RFID(byte chipSelectPin, byte resetPowerDownPin, Lock *_lock, bool _enabled = true);
-        virtual ~RFID() { this->rfid.PCD_SoftPowerDown(); }
+        /*
+            @param reset not connected by default
+        */
+        RFID(byte irq, byte reset, Lock *_lock, bool _enabled = true);
+        virtual ~RFID() {}
 
-        void begin();
+        /*
+            @return true if the RFID-PN532 was setup successfully - else false
+        */
+        bool begin();
 
         /*
             read for a tag id
-            @param async false=wait until a tag is present. true=read 1 time and return imidiately
-            @return the tag id if a tag was present else an empty string
+            @return the tag id if a tag was present else an empty UID-object
         */
-        UID read_Tag_UID(bool async = true);
+        UID read_Tag_UID();
 
         /*
             reads from the RFID_module for a tag and adds it to database at id id
@@ -56,7 +62,7 @@ namespace RFID
             removes tag from the tag with uidid tag_uid from the allowed_tags
             @return true if one or more tags where removed - else false
         */
-        bool remove_tag(UID tag_uid);
+        bool remove_tag(const UID &tag_uid);
         /*
             removes all saved tags from the allowed_tags_database
         */
@@ -76,272 +82,14 @@ namespace RFID
         bool id_used(unsigned short id) const;
         // bool in_database(String uid);
 
+        /*
+            read for uids and return a succes if a uid was read that matches to one of the database
+        */
         unlock_authentication_reports read() override;
 
     private:
-        MFRC522 rfid;
+        Adafruit_PN532 rfid;
         UID allowed_tags[NUM_OF_TAGS];
     };
 
 } // namespace RFID
-
-// ------------ Implementations ------------
-
-RFID::RFID::RFID(byte chipSelectPin, byte resetPowerDownPin,
-                 Lock *_lock, bool _enabled) : Unlock_Object(_lock, _enabled), rfid(chipSelectPin, resetPowerDownPin)
-{
-    for (unsigned short i = 0; i < NUM_OF_TAGS; ++i)
-    {
-        this->allowed_tags[i].clear();
-    }
-}
-
-void RFID::RFID::begin()
-{
-    SPI.begin();                          // Arduino interface which is necessarily for RFID(SPI is a global variable from Arduino)
-    this->rfid.PCD_Init();                // starting and initialising rfid
-    this->rfid.PCD_DumpVersionToSerial(); // printing RFID_Version to serial
-    if (this->rfid.PCD_PerformSelfTest())
-    {
-        logger.log(F("RFID:passed self test"), Log::log_level::L_INFO);
-    }
-    else
-    {
-        logger.log(F("RFID: Self test failed"), Log::log_level::L_WARNING);
-    }
-    logger.log(F("RFID: status ready..."), Log::log_level::L_INFO);
-}
-
-Unlock_Object::unlock_authentication_reports RFID::RFID::read()
-{
-    if (!this->is_enabled())
-    {
-        return Unlock_Object::unlock_authentication_reports::UNLOCK_OBJECT_DISABLED;
-    }
-
-    UID tag_uid;
-    if (this->rfid.PICC_IsNewCardPresent())
-    {
-        if (this->rfid.PICC_ReadCardSerial()) // successfully read the card UID
-        {
-            tag_uid = this->rfid.uid;
-
-            DEBUG_PRINT(F("Read tag with UID: "));
-            DEBUG_PRINTLN(tag_uid.to_string());
-
-            // iterate over the stored allowed tags and check for a matching tag
-            for (unsigned short i = 0; i < NUM_OF_TAGS; ++i)
-            {
-                if (this->allowed_tags[i] == tag_uid)
-                {
-                    return Unlock_Object::unlock_authentication_reports::AUTHORIZED_UNLOCK_OBJECT;
-                }
-            }
-        }
-        else // error while reading the card UID
-        {
-            return Unlock_Object::unlock_authentication_reports::UNLOCK_OBJECT_READ_ERROR;
-        }
-    }
-    else
-    {
-        return Unlock_Object::unlock_authentication_reports::NO_UNLOCK_OBJECT_PRESENT;
-    }
-
-    return Unlock_Object::unlock_authentication_reports::UNAUTHORIZED_UNLOCK_OBJECT;
-}
-
-bool RFID::RFID::id_used(unsigned short id) const
-{
-    if (id > NUM_OF_TAGS - 1)
-    {
-        DEBUG_PRINT(F("id is out of range"))
-        // throw an error
-    }
-    return this->allowed_tags[id].is_set();
-}
-
-bool RFID::RFID::read_add_tag(unsigned short id)
-{
-    if (id > NUM_OF_TAGS - 1)
-    {
-        DEBUG_PRINTLN(F("id is out of range"));
-        // throw an error
-    }
-    UID tag_uid = this->read_Tag_UID();
-    if (tag_uid)
-    {
-        this->allowed_tags[id] = tag_uid;
-        return true;
-    }
-    return false;
-}
-
-void RFID::RFID::add_tag(unsigned short id, UID tag_uid)
-{
-    if (id > NUM_OF_TAGS - 1)
-    {
-        // throw an error
-    }
-    this->allowed_tags[id] = tag_uid;
-}
-
-void RFID::RFID::remove_tag(unsigned short id)
-{
-    if (id > NUM_OF_TAGS - 1)
-    {
-        DEBUG_PRINTLN(F("id is out of range"));
-        // throw an error
-    }
-    this->allowed_tags[id].clear();
-}
-
-bool RFID::RFID::remove_tag(UID tag_uid)
-{
-    bool least_one_tag_removed = false;
-    for (unsigned short i = 0; i < NUM_OF_TAGS; ++i)
-    {
-        if (this->allowed_tags[i] == tag_uid)
-        {
-            this->allowed_tags[i].clear();
-            least_one_tag_removed = true;
-        }
-    }
-    return least_one_tag_removed;
-}
-
-void RFID::RFID::clear_database()
-{
-    for (unsigned short i = 0; i < NUM_OF_TAGS; ++i)
-    {
-        this->allowed_tags[i].clear();
-    }
-}
-
-RFID::UID &RFID::RFID::get_tag_uid(unsigned short id)
-{
-    if (id > NUM_OF_TAGS - 1)
-    {
-        DEBUG_PRINTLN(F("id is out of range"));
-        // return;
-        // throw an error
-    }
-
-    if (!this->allowed_tags[id])
-    {
-        DEBUG_PRINTLN(F("tag isnt set"));
-        // return;
-        // throw error tag isnt set...
-    }
-    else
-    {
-        // return this->allowed_tags[id].operator bool();
-        return this->allowed_tags[id];
-    }
-}
-
-int RFID::RFID::get_tag_id(UID tag_uid)
-{
-    for (unsigned short i = 0; i < NUM_OF_TAGS; ++i)
-    {
-        if (this->allowed_tags[i] == tag_uid)
-        {
-            return i;
-        }
-    }
-    return -1;
-}
-
-RFID::UID RFID::RFID::read_Tag_UID(bool async)
-{
-    UID uid;
-    String uid_str;
-    if (!async) // if we dont run in async mode wait until an new card is present
-    {
-        while (!this->rfid.PICC_IsNewCardPresent()) // returns true if a card is present
-            delay(10);
-    }
-    else
-    {
-        if (!this->rfid.PICC_IsNewCardPresent()) // if no card is "online"
-        {
-            auto err_code = rfid.PICC_HaltA(); // halt the reader in order to not read the same card again and again
-            if (err_code != MFRC522::StatusCode::STATUS_OK)
-            {
-                String error_msg(F("RFID: error reading tag - halting - error_message: "));
-                error_msg += error_message(err_code);
-                logger.log(error_msg, Log::log_level::L_ERROR);
-                DEBUG_PRINTLN(error_msg);
-            }
-            return UID(); // return a empty uid
-        }
-    }
-
-    // a new card is present...
-    if (rfid.PICC_ReadCardSerial())
-    { // starting read the card
-        uid = this->rfid.uid;
-    }
-    rfid.PICC_HaltA(); // halt the reader in order to not read the same card again and again
-    return uid;
-}
-
-const char *RFID::error_message(MFRC522::StatusCode error_code)
-{
-    const char *err_msg = nullptr;
-    switch (error_code)
-    {
-    case MFRC522::StatusCode::STATUS_COLLISION:
-    {
-        err_msg = "Collission detected";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_CRC_WRONG:
-
-    {
-        err_msg = "The CRC_A does not match";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_ERROR:
-    {
-        err_msg = "Error in communication";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_INTERNAL_ERROR:
-    {
-        err_msg = "Internal error in the code(in RFID-library). Should not happen ;-)";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_INVALID:
-    {
-        err_msg = "Invalid argument.";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_MIFARE_NACK:
-    {
-        err_msg = "A MIFARE PICC responded with NAK.";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_NO_ROOM:
-    {
-        err_msg = "A buffer is not big enough.";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_OK:
-    {
-        err_msg = "Success";
-        break;
-    }
-    case MFRC522::StatusCode::STATUS_TIMEOUT:
-    {
-        err_msg = "Timeout in communication.";
-        break;
-    }
-    default:
-    {
-        err_msg = "Unknown ERROR";
-        break;
-    }
-    }
-    return err_msg;
-}
